requirements.txt
annotated-types==0.6.0
blis==0.7.11
build==1.0.3
CacheControl==0.13.1
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.2.0
crashtest==0.4.1
cycler==0.12.1
cymem==2.0.8
distlib==0.3.7
dulwich==0.21.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
fastjsonschema==2.19.0
filelock==3.13.1
fonttools==4.45.1
hashin==0.17.0
idna==3.6
importlib-metadata==6.8.0
installer==0.7.0
jaraco.classes==3.3.0
Jinja2==3.1.2
joblib==1.3.2
keyring==24.3.0
kiwisolver==1.4.5
langcodes==3.3.0
load-dotenv==0.1.0
MarkupSafe==2.1.3
matplotlib==3.8.2
more-itertools==10.1.0
msgpack==1.0.7
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
pexpect==4.9.0
Pillow==10.1.0
pip-api==0.0.30
pkginfo==1.9.6
platformdirs==3.11.0
poetry-core==1.8.1
poetry-plugin-export==1.6.0
praw==7.7.1
prawcore==2.4.0
preshed==3.0.9
ptyprocess==0.7.0
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32-ctypes==0.2.2
rapidfuzz==3.5.2
regex==2023.10.3
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.1
tomlkit==0.12.3
tqdm==4.66.1
trove-classifiers==2023.11.29
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
update-checker==0.18.0
urllib3==2.1.0
virtualenv==20.24.7
wasabi==1.1.2
weasel==0.3.4
websocket-client==1.6.4
wordcloud==1.9.2
zipp==3.17.0
